+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### 2024-10-02
 
 - Added file source.
+- Changed fake source to stream with 720p resolution.
+- Added black and white filter as an option.
+- Added timestamp (option).
 
 ### 2024-10-01